pyproject.toml
[project]
name = "comfyui-levelpixel-advanced"
description = "Advanced nodes of the [Level Pixel](https://github.com/LevelPixel/ComfyUI-LevelPixel-Advanced) company (aka levelpixel, LP). Includes convenient advanced nodes for working with LLM и VLM models (LLaVa, Multimodal Generator) with GGUF format. Qwen2.5-VL and Qwen2.5 supported, based on llama-mtmd. Also included is a node for the RAM model. Nodes have the ability to automatically unload models from VRAM. This package also includes Autotagger (WD), Image Remove Background (based on RMBG-2.0, on BiRefNet, on BiRefNet-HR-Matter, on RemBG)."
version = "1.2.3"
license = { file = "LICENSE" }
dependencies = [
    "torch>=2.0.1", 
    "torchvision>=0.15.2", 
    "pillow>=9.4.0", 
    "numpy>=1.26.4", 
    "huggingface-hub>=0.19.0", 
    "transformers>=4.35.0", 
    "safetensors>=0.3.0",
    "scikit-build-core>=0.10.7",
    "transparent-background>=1.2.4",
    "timm>=0.4.12",
    "tqdm>=4.65.0",
    "fairscale>=0.4.4",
    "rembg>=2.0.66" ]

[project.urls]
Repository = "https://github.com/LevelPixel/ComfyUI-LevelPixel-Advanced"

[tool.comfy]
PublisherId = "levelpixel"
DisplayName = "ComfyUI Level Pixel Advanced"
Icon = ""